requirements.txt
matplotlib==3.3.4
plotly==4.7.0
streamlit==0.79.0
pandas==1.2.3
numpy==1.19.5
xgboost==0.90
Pillow==8.3.1
scikit_learn==0.24.2
